* Added support of build for 32-bit and 64-bit RISC-V architecture (contributed by Xeonacid at GitHub).

+++ src/plugins/art_delay.cpp
@@ -577,7 +577,7 @@ namespace lsp
 
     size_t art_delay_base::decode_max_delay_value(size_t v)
     {
-        float seconds = (v < sizeof(art_delay_max)/sizeof(uint8_t)) ? art_delay_max[v] : 1.0f;
+        float seconds = (v < sizeof(art_delay_max)/sizeof(uint16_t)) ? art_delay_max[v] : 1.0f;
         return seconds_to_samples(fSampleRate, seconds);
     }
